+++ backend/src/application/services/candidateService.ts
@@ -1,65 +1,41 @@
 import { Candidate } from '../../domain/models/Candidate';
 import { validateCandidateData } from '../validator';
-import { Education } from '../../domain/models/Education';
-import { WorkExperience } from '../../domain/models/WorkExperience';
-import { Resume } from '../../domain/models/Resume';
+import { ICandidateRepository } from '../../domain/repositories/ICandidateRepository';
 
-export const addCandidate = async (candidateData: any) => {
-    try {
-        validateCandidateData(candidateData); // Validar los datos del candidato
-    } catch (error: any) {
-        throw new Error(error);
+export class CandidateService {
+
+    private candidateRepository: ICandidateRepository;
+
+    constructor(candidateRepository: ICandidateRepository) {
+        this.candidateRepository = candidateRepository;
     }
 
-    const candidate = new Candidate(candidateData); // Crear una instancia del modelo Candidate
-    try {
-        const savedCandidate = await candidate.save(); // Guardar el candidato en la base de datos
-        const candidateId = savedCandidate.id; // Obtener el ID del candidato guardado
+    async addCandidate(candidateData: any): Promise<Candidate> {
+        try {
+            validateCandidateData(candidateData); // Validate candidate data
+        } catch (error: any) {
+            throw new Error(error);
+        }
 
-        // Guardar la educación del candidato
-        if (candidateData.educations) {
-            for (const education of candidateData.educations) {
-                const educationModel = new Education(education);
-                educationModel.candidateId = candidateId;
-                await educationModel.save();
-                candidate.education.push(educationModel);
+        const candidate = new Candidate(candidateData); // Create a new Candidate instance
+        try {
+            return await this.candidateRepository.save(candidate); // Save the candidate to the database
+        } catch (error: any) {
+            if (error.code === 'P2002') {
+                // Unique constraint failed on the fields: (`email`)
+                throw new Error('The email already exists');
+            } else {
+                throw error;
             }
         }
+    }
 
-        // Guardar la experiencia laboral del candidato
-        if (candidateData.workExperiences) {
-            for (const experience of candidateData.workExperiences) {
-                const experienceModel = new WorkExperience(experience);
-                experienceModel.candidateId = candidateId;
-                await experienceModel.save();
-                candidate.workExperience.push(experienceModel);
-            }
-        }
-
-        // Guardar los archivos de CV
-        if (candidateData.cv && Object.keys(candidateData.cv).length > 0) {
-            const resumeModel = new Resume(candidateData.cv);
-            resumeModel.candidateId = candidateId;
-            await resumeModel.save();
-            candidate.resumes.push(resumeModel);
-        }
-        return savedCandidate;
-    } catch (error: any) {
-        if (error.code === 'P2002') {
-            // Unique constraint failed on the fields: (`email`)
-            throw new Error('The email already exists in the database');
-        } else {
-            throw error;
+    async findCandidateById(id: number): Promise<Candidate | null> {
+        try {
+            return await this.candidateRepository.findById(id);
+        } catch (error) {
+            console.error('Error finding candidate:', error);
+            throw new Error('Error retrieving candidate');
         }
     }
-};
-
-export const findCandidateById = async (id: number): Promise<Candidate | null> => {
-    try {
-        const candidate = await Candidate.findOne(id); // Cambio aquí: pasar directamente el id
-        return candidate;
-    } catch (error) {
-        console.error('Error al buscar el candidato:', error);
-        throw new Error('Error al recuperar el candidato');
-    }
-};
+}
