+++ backend/src/presentation/controllers/candidateController.ts
@@ -1,16 +1,19 @@
 import { Request, Response } from 'express';
-import { addCandidate, findCandidateById } from '../../application/services/candidateService';
+import { CandidateService } from '../../application/services/candidateService';
+import { CandidateRepository } from '../../infrastructure/repositories/CandidateRepository';
 
-export const addCandidateController = async (req: Request, res: Response) => {
+const candidateRepository = new CandidateRepository();
+const candidateService = new CandidateService(candidateRepository);
+
+export const addCandidate = async (req: Request, res: Response) => {
     try {
-        const candidateData = req.body;
-        const candidate = await addCandidate(candidateData);
+        const candidate = await candidateService.addCandidate(req.body);
         res.status(201).json({ message: 'Candidate added successfully', data: candidate });
     } catch (error: unknown) {
         if (error instanceof Error) {
             res.status(400).json({ message: 'Error adding candidate', error: error.message });
         } else {
-            res.status(400).json({ message: 'Error adding candidate', error: 'Unknown error' });
+            res.status(500).json({ error: 'Internal Server Error' });
         }
     }
 };
@@ -21,7 +24,7 @@ export const getCandidateById = async (req: Request, res: Response) => {
         if (isNaN(id)) {
             return res.status(400).json({ error: 'Invalid ID format' });
         }
-        const candidate = await findCandidateById(id);
+        const candidate = await candidateService.findCandidateById(id);
         if (!candidate) {
             return res.status(404).json({ error: 'Candidate not found' });
         }
@@ -30,5 +33,3 @@ export const getCandidateById = async (req: Request, res: Response) => {
         res.status(500).json({ error: 'Internal Server Error' });
     }
 };
-
-export { addCandidate };
