+++ prompts/prompts-OGM.md
@@ -1,0 +1,16 @@
+@workspace can you give me recommendations on how to better use SOLID and DDD practices within the backend module of this project? We will implement some of your answers.
+
+---
+
+Let's implement the CandidateRepository class you suggested as a DDD best practice. Let's also incorporate the Single Responsibility and Dependency Inversion principles from SOLID to move the corresponding logic away from the #file:Candidate.ts class.
+
+---
+
+Update #file:candidateService.ts to use the newly added #file:CandidateRepository.ts to update #file:Candidate.ts instances.
+
+---
+
+There's a lot of duplicated logic between the #file:CandidateService.ts class and #file:CandidateRepository.ts classes, can you clean it up?
+
+---
+
